requirements.txt
beautifulsoup4==4.12.2
notion-client==2.0.0
python-dotenv==1.0.0